% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Print "Fars Map State"}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer (or integer-coercible) for state.}

\item{year}{integer (or integer-coercible) 4-digit year.}
}
\description{
This is a function that takes \code{state.num} and \code{year} arguments
and maps the locations of rows (accidents) within the referenced state.
}
\details{
Errors could come from years being improperly formatted or state.num not
being coercible to integer.
}
\examples{
\dontrun{
fars_map_state(13, 2013)
}

}
