% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read}
\alias{fars_read}
\title{Print "Fars Read"}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file to be read}
}
\value{
This function returns the data in user-provided \code{filename}
   as a tbl.
}
\description{
This is a simple function that takes a filename (provided by user
with \code{filename} argument), checks if file exists and if so reads
the files contents as a csv and outputs the data as a tbl. Errors could
result from the file being a format the cannot be read by readr::read_csv
}
\examples{
fars_read('myfile.csv')

}
