% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Print "Fars Read Years"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list (or vector) of years}
}
\value{
a list with items for each inputed year, with month and year
from the file for all years with corresponding files and NULLs for all
years without files.
}
\description{
This is a function that takes a vector or list of years (provided by user
with \code{years} argument), and outputs a list of years each containing a
dataframe of MONTHS and year, for all rows in the file with the input year.
}
\details{
Errors could come from years being improperly formatted.
}
\examples{
\dontrun{
fars_read_years(list('2013','2015'))
}

}
