% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Print "Fars Summarize Years"}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list (or vector) of years}
}
\value{
dataframe with columns for each user inputed year that had an associated
file, and rows (accidents) for each MONTH, containing counts of rows for each month/year
across files.
}
\description{
This is a function that takes a vector or list of years (provided by user
with \code{years} argument), and outputs a dataframe with columns for each
year and rows for each month containing counts of rows (accidents) in each
year's files for each month.
}
\details{
Errors could come from years being improperly formatted.
}
\examples{
fars_summarize_years(list('2002','2007'))

}
