% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_filename}
\alias{make_filename}
\title{Print "Make Filename"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or numeric year.}
}
\value{
This function returns a string of the file name for the
   given year.
}
\description{
This is a simple function that takes a year (provided by user
with \code{year} argument), and returns a character string filename
of the format "accident_[YEAR].csv.bz2". Errors could result from
the year being given in a format that cannot be coerced to integer.
}
\examples{
\dontrun{
make_filename('2013')
}

}
